requirements.txt
asgiref==3.3.1
Django==3.1.3
django-phone-field==1.8.1
mysqlclient==2.0.1
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
